backend/requirements.txt
Django==3.1.1
djangorestframework==3.11.1
django-filter
